pyproject.toml
[project]
name = "doc-classifier-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.13.0",
    "boto3>=1.34.0",
    "casbin>=1.41.0",
    "casbin-sqlalchemy-adapter>=0.6.0",
    "fastapi>=0.115.0",
    "fastapi-cache2[redis]>=0.2.0",
    "fastapi-users[sqlalchemy]>=13.0.0",
    "httpx>=0.27.0",
    "hvac>=2.1.0",
    "mypy>=1.9.0",
    "paramiko>=3.4.0",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=10.0.0",
    "pre-commit>=3.7.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "redis>=4.2.0",
    "requests>=2.31.0",
    "rq>=1.16.0",
    "ruff>=0.4.0",
    "sqlalchemy>=2.0.0",
    "structlog>=24.0.0",
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "uvicorn[standard]>=0.30.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
